Initial breaking apart of admin view into the two different tabs
