+++ admin/src/FontAwesomeAdminView.module.css
@@ -18,7 +18,7 @@
 }
 
 .tab-header button:disabled {
-    cursor: not-allowed;
+    cursor: default;
     border-bottom: 4px solid darkgray;
     color: unset;
 }
