+++ admin/src/SettingsTab.js
@@ -1,7 +1,8 @@
 import React from 'react'
+import Options from './Options'
 
 export default function SettingsTab() {
   return <div>
-    settings tab
+    <Options />
   </div>
 }
